update 0.2.0 - Implementation for `v1` condition-status endpoints - Generic response creation - reworked condition implementation (db + rest functions)

+++ internal/rest/conditions.go
@@ -12,30 +12,34 @@ import (
 func (r *RestService) GetConditions(c *gin.Context) {
 	res, err := database.SelectConditions(r.DB)
 	if err != nil {
-		panic(err)
+		createErrorResponse(c, http.StatusInternalServerError, "Error on database connection.")
+		return
 	}
-	c.IndentedJSON(http.StatusOK, res)
+	createDataResponse(c, http.StatusOK, res)
 }
 
 func (r *RestService) PostCondition(c *gin.Context) {
 	var newCondition models.Condition
 	if err := c.BindJSON(&newCondition); err != nil {
+		createErrorResponse(c, http.StatusBadRequest, "Invalid json data.")
 		fmt.Println(err)
 		return
 	}
 	newCondition.Creator = "SYSTEM"
 	err := database.InsertConditionsSpecific(r.DB, &newCondition)
 	if err != nil {
+		createErrorResponse(c, http.StatusInternalServerError, "Failed to save new condition.")
 		fmt.Println(err)
 		return
 	}
 	newConditionStatus := models.ConditionStatus{ID: newCondition.ID, Status: false, LastChangedBy: "SYSTEM", LastChangedAt: time.Now()}
 	err = database.InsertConditionStatusSpecific(r.DB, &newConditionStatus)
 	if err != nil {
+		createErrorResponse(c, http.StatusInternalServerError, "Failed to save new condition status.")
 		fmt.Println(err)
 		return
 	}
-	c.IndentedJSON(http.StatusCreated, newCondition)
+	createDataResponse(c, http.StatusCreated, newCondition)
 }
 
 func (r *RestService) getConditionByID(id string) (*models.Condition, error) {
@@ -47,10 +51,14 @@ func (r *RestService) GetConditionSpecific(c *gin.Context) {
 	condition, err := r.getConditionByID(id)
 	if err != nil {
 		fmt.Println(err)
-		c.IndentedJSON(http.StatusNotFound, gin.H{"message": "condition not found"})
+		createErrorResponse(c, http.StatusInternalServerError, fmt.Sprintf("Search for Condition with ID %s failed.", id))
 		return
 	}
-	c.IndentedJSON(http.StatusOK, condition)
+	if condition == nil {
+		createErrorResponse(c, http.StatusNotFound, fmt.Sprintf("Condition with ID %s not found.", id))
+		return
+	}
+	createDataResponse(c, http.StatusOK, condition)
 }
 
 func (r *RestService) DeleteConditionSpecific(c *gin.Context) {
@@ -61,28 +69,36 @@ func (r *RestService) DeleteConditionSpecific(c *gin.Context) {
 		c.IndentedJSON(http.StatusNotFound, gin.H{"message": "condition not found"})
 		return
 	}
+	if condition == nil {
+		createErrorResponse(c, http.StatusNotFound, fmt.Sprintf("Condition with ID %s not found.", id))
+		return
+	}
 	err = database.DeleteConditionsSpecific(r.DB, id)
 	if err != nil {
-		panic(err)
+		createErrorResponse(c, http.StatusInternalServerError, fmt.Sprintf("Deletion of Condition with ID %s failed.", id))
+		fmt.Println(err)
 	}
 
 	err = database.DeleteConditionStatusSpecific(r.DB, id)
 	if err != nil {
-		panic(err)
+		createErrorResponse(c, http.StatusInternalServerError, fmt.Sprintf("Deletion of Condition Status with ID %s failed.", id))
+		fmt.Println(err)
 	}
-	c.IndentedJSON(http.StatusOK, condition)
+	createDataResponse(c, http.StatusOK, condition)
 }
 
 func (r *RestService) UpdateCondition(c *gin.Context) {
 	var newCondition models.Condition
 	if err := c.BindJSON(&newCondition); err != nil {
+		createErrorResponse(c, http.StatusBadRequest, "Invalid json data.")
 		fmt.Println(err)
 		return
 	}
 	err := database.UpdateConditionsSpecific(r.DB, &newCondition)
 	if err != nil {
+		createErrorResponse(c, http.StatusInternalServerError, fmt.Sprintf("Update of Condition with ID %s failed.", newCondition.ID))
 		fmt.Println(err)
 		return
 	}
-	c.IndentedJSON(http.StatusOK, newCondition)
+	createDataResponse(c, http.StatusOK, newCondition)
 }
